Fix a few more printf errors

+++ tagutils/tagutils-misc.c
@@ -190,7 +190,7 @@ vc_scan(struct song_metadata *psong, const char *comment, const size_t length)
 	{
 		if( strncasecmp(comment, "LYRICS=", 7) != 0 )
 		{
-			DPRINTF(E_WARN, L_SCANNER, "Vorbis %.*s too long [%s]\n", (index(comment, '=')-comment), comment, psong->path);
+			DPRINTF(E_WARN, L_SCANNER, "Vorbis %.*s too long [%s]\n", (int)(index(comment, '=')-comment), comment, psong->path);
 		}
 		return;
 	}
